Fix false positive introduced for foreach reusing variable name
Fixes #2861
