+++ tests/files/expected/0705_undeclared_class.php.expected
@@ -1,0 +1,2 @@
+%s:6 PhanTypeMismatchArgumentInternal Argument 1 ($numerator) is \ReflectionFunctionAbstract|\ReflectionMethod|\Reflector but \intdiv() takes int
+%s:9 PhanUndeclaredMethod Call to undeclared method \ReflectionClass::getMissingMethod
